Revert "Db entity mechanics (#26)"
This reverts commit eacae54739f93d93707e37c5f742156d24e0d4a3.
