+++ .gitignore
@@ -24,7 +24,6 @@ yarn-error.log*
 
 .env
 
-
 StorySquad-2922ddf6e57d.json*
 readability.json*
 scripts.spec.ts*
