--- Pipfile
@@ -1,22 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pylint = "*"
-flake8 = "*"
-
-[packages]
-google-cloud-datastore = "*"
-google-cloud-vision = "*"
-python-decouple = "*"
-python-dotenv = "*"
-pandas = "*"
-pyinstaller = "*"
-python-datauri = "*"
-textstat = "*"
-scikit-learn = "*"
-
-[requires]
-python_version = "3.7"
