--- src/database/entity/Deprecated/DueDates.ts
@@ -1,25 +1,0 @@
-// import { Column } from 'typeorm';
-
-// import { Activities } from '../../models';
-// // currently implements syntax isn't actually being used
-// class DueDates implements Activities {
-//     @Column()
-//     reading: Date;
-
-//     @Column()
-//     writing: Date;
-
-//     @Column()
-//     submission: Date;
-
-//     // @Column()
-//     // teamReview: Date;
-
-//     // @Column()
-//     // randomReview: Date;
-
-//     // @Column()
-//     // results: Date;
-// }
-
-// export { DueDates };
