--- src/database/entity/User/index.ts
@@ -1,4 +1,0 @@
-export * from './Child'
-export * from './Admin'
-export * from './Parent'
-export * from './Preferences'
