+++ src/database/entity/index.ts
@@ -1,17 +1,6 @@
-//Feedback
-export * from './Feedback/Child_Votes';
-export * from './Feedback/Feedback';
-//Matching
-export * from './Matching/Matches';
-//Story
-export * from './Story/Canon';
-export * from './Story/Round';
-export * from './Story/Cohort';
-export * from './Story/Story';
-//Submission
-export * from './Submission/Submission';
-export * from './Submission/Readability';
-//User
-export * from './User/Admin';
-export * from './User/Child';
-export * from './User/Parent';
+export * from './Admin';
+export * from './Canon';
+export * from './Child';
+export * from './Cohort';
+export * from './Parent';
+export * from './Submissions';
