+++ src/routes/canon/canon.routes.spec.ts
@@ -2,8 +2,7 @@ import * as request from 'supertest';
 import * as express from 'express';
 import { plainToClass } from 'class-transformer';
 
-import { Admin } from '../../database/entity/User';
-import { Canon } from '../../database/entity/Story'
+import { Admin, Canon } from '../../database/entity';
 
 import { canonRoutes } from './canon.routes';
