+++ src/routes/child/child.routes.ts
@@ -6,7 +6,7 @@ import { Parent, Child } from '../../database/entity';
 import { Only } from '../../middleware';
 import { connection } from '../../util/typeorm-connection';
 
-import { Cohort } from '../../database/entity/Story/Cohort';
+import { Cohort } from '../../database/entity/Cohort';
 
 const childRoutes = Router();
