+++ src/routes/cohort/cohort.routes.ts
@@ -2,8 +2,7 @@ import { Router } from 'express';
 import { getRepository } from 'typeorm';
 
 import { connection } from '../../util/typeorm-connection';
-import { Cohort } from '../../database/entity/Story';
-import { Child, Admin } from '../../database/entity/User'
+import { Child, Admin, Cohort } from '../../database/entity';
 import { Only } from '../../middleware';
 
 const cohortRoutes = Router();
