+++ src/util/scripts/readability.py
@@ -1,8 +1,8 @@
 from sys import stdin, stdout
 from json import loads, dumps
 from decouple import config
-#from textstat.textstat import textstat
 import textstat
+
 import urllib.request
 
 def get_stats(text):
